requirements.txt
thordata-sdk>=1.4.0
mcp[cli]
html2text
python-dotenv
markdownify
pydantic
pydantic-settings
httpx